feat: add radio field; wrap inputs with label
moved toward a strategy of
```html
<label>
  Label
  <input type="text" name="foo" />
</label>
```

this strategy makes the input accessible without requiring an ID.
i want to avoid usage of IDs because i can't guarantee that the name of the input
is a unique ID.
this would mean the user would have to pass in IDs which would blow up the API,
especially when you'd have to do it for overrides too.

the downside of this approach is most devs probably don't realize this wrapped labeled
syntax is valid (partially b/c most devs aren't super familiar with a11y).

needing to specify label became pronounced with the radio field since it will have multiple
labels with diff `htmlFor`.
the wrapped strategy makes it easier to specify how each value is labeled without needing IDs